Add code to allow ZapRedraw to plot directly from 1bpp bitmaps in 32bpp modes. Aside from the obvious memory savings, this has also resulted in an 8% speed increase due to the reduced bandwidth.
This will also make it easier to add transparency, since doing it with
the existing code involves reading screen memory which is appalingly
slow on Iyonix (it took over two seconds to redraw a smallish window).

Fixed a couple of other bugs in DSAVDU mode which I introduced with some
earlier optimisations.