Add support for multiple zap font resources.
Ideally we'd use *ZapFontInstall to do this, but unlike the Font Manager
we can't guarantee ZapRedraw will be loaded (in fact at boot time it
definitely won't). Instead use a BASIC program to install the path and make
sure there aren't multiple copies of the same resource application, and get
ZapRedraw to rescan the path on the off chance that it _is_ loaded.